Cross platform supposrt for mobile and pc

+++ lua/core/keymaps.lua
@@ -62,41 +62,44 @@ vim.keymap.set("v", "<C-j>", ":m '>+1<CR>gv=gv", { desc = "Move selection down" 
 vim.keymap.set("v", "<C-k>", ":m '<-2<CR>gv=gv", { desc = "Move selection up" })
 
 
--- Emoji Picker Integration
+local utils = require("core.utils")
 
-vim.keymap.set({ "n", "i" }, "<F2>", function()
-    local buf = vim.api.nvim_create_buf(false, true)
-    -- Compact size matching the emo script
-    local width = 46
-    local height = 12
-    local win = vim.api.nvim_open_win(buf, true, {
-        relative = "editor",
-        width = width,
-        height = height,
-        col = (vim.o.columns - width) / 2,
-        row = (vim.o.lines - height) / 2,
-        style = "minimal",
-        border = "none" -- We let emo draw the border
-    })
-    local tmp = os.tmpname()
-    -- Pass EMO_EMBED=1 to remove internal margins
-    vim.cmd("terminal EMO_EMBED=1 ~/.local/bin/emo > " .. tmp)
-    vim.bo.bufhidden = "wipe"
-    vim.api.nvim_create_autocmd("TermClose", {
-        buffer = 0,
-        once = true,
-        callback = function()
-            if vim.api.nvim_win_is_valid(win) then vim.api.nvim_win_close(win, true) end
-            local f = io.open(tmp, "r")
-            if f then
-                local emoji = f:read("*a")
-                f:close()
-                os.remove(tmp)
-                if emoji and emoji ~= "" then
-                    vim.api.nvim_put({ emoji }, "c", false, true)
+-- Emoji Picker Integration (Termux Only)
+if utils.is_android then
+    vim.keymap.set({ "n", "i" }, "<F2>", function()
+        local buf = vim.api.nvim_create_buf(false, true)
+        -- Compact size matching the emo script
+        local width = 46
+        local height = 12
+        local win = vim.api.nvim_open_win(buf, true, {
+            relative = "editor",
+            width = width,
+            height = height,
+            col = (vim.o.columns - width) / 2,
+            row = (vim.o.lines - height) / 2,
+            style = "minimal",
+            border = "none" -- We let emo draw the border
+        })
+        local tmp = os.tmpname()
+        -- Pass EMO_EMBED=1 to remove internal margins
+        vim.cmd("terminal EMO_EMBED=1 ~/.local/bin/emo > " .. tmp)
+        vim.bo.bufhidden = "wipe"
+        vim.api.nvim_create_autocmd("TermClose", {
+            buffer = 0,
+            once = true,
+            callback = function()
+                if vim.api.nvim_win_is_valid(win) then vim.api.nvim_win_close(win, true) end
+                local f = io.open(tmp, "r")
+                if f then
+                    local emoji = f:read("*a")
+                    f:close()
+                    os.remove(tmp)
+                    if emoji and emoji ~= "" then
+                        vim.api.nvim_put({ emoji }, "c", false, true)
+                    end
                 end
             end
-        end
-    })
-    vim.cmd("startinsert")
-end, { desc = "Termux Emoji Picker" })
+        })
+        vim.cmd("startinsert")
+    end, { desc = "Termux Emoji Picker" })
+end
